personal-ai-system/requirements.txt
# Core AI/LLM Libraries
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
langchain-google-genai==2.0.5
langgraph==0.2.45
langsmith==0.2.3

# Agent Frameworks
autogen-agentchat==0.4.0
crewai==0.86.0

# Vector Database & Embeddings
qdrant-client==1.12.1
sentence-transformers==3.3.1

# Caching & Message Queue
redis==5.2.0
celery==5.4.0

# Web & API
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
requests==2.32.3
beautifulsoup4==4.12.3

# CLI Interface
rich==13.9.4
typer==0.15.1
prompt-toolkit==3.0.48

# Data Processing
pandas==2.2.3
numpy==2.1.3
python-dateutil==2.9.0

# Email & Calendar
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
exchangelib==5.4.3

# Task Management Integrations
jira==3.8.0
asana==5.0.10
todoist-api-python==2.1.8

# Web Scraping & Search
playwright==1.48.0
duckduckgo-search==7.0.1
googlesearch-python==1.2.5

# Supabase Integration
supabase==2.10.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
pyyaml==6.0.2
loguru==0.7.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Monitoring & Observability
prometheus-client==0.21.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
